app/oauth2.py
from jose import jwt, JWTError
from datetime import datetime, timedelta

SECRET_KEY = "330bcb19c85f952d19f269902ca3ba793a2c869618c3492022ce79ae7c1f1468"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30


def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.now() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})

    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt
