requirements.txt
anyio==3.6.1
asgiref==3.5.2
bcrypt==3.2.2
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
orjson==3.7.6
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
